Adjusted the versions expander
